refactor(setup): extract wizard step 4 (backup)
